# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Clipup Project Management & Messaging API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spaces:
    get:
      operationId: listSpaces
      tags: [spaces]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Space"
    post:
      operationId: createSpace
      tags: [spaces]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
  /spaces/{id}:
    get:
      operationId: getSpace
      tags: [spaces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
    patch:
      operationId: updateSpace
      tags: [spaces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
    delete:
      operationId: deleteSpace
      tags: [spaces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PROJECTS ───────────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: dueToday
          in: query
          schema:
            type: boolean
        - name: overdue
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/comments:
    get:
      operationId: getTaskComments
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: createTaskComment
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskCommentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"
  /tasks/{id}/subtasks:
    get:
      operationId: getSubtasks
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks/{id}/watchers:
    post:
      operationId: addTaskWatcher
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWatcherBody"
      responses:
        "200":
          description: OK

  # ─── GOALS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Space:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        icon:
          type: string
        ownerId:
          type: integer
        ownerName:
          type: string
        projectCount:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, color, icon, ownerId, ownerName, projectCount, memberCount, createdAt]
    CreateSpaceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        icon:
          type: string
        ownerId:
          type: integer
      required: [name, color, icon, ownerId]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        parentTaskId:
          type: ["integer", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        assigneeAvatar:
          type: ["string", "null"]
        creatorId:
          type: integer
        creatorName:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date-time
        startDate:
          type: ["string", "null"]
          format: date-time
        estimatedHours:
          type: ["number", "null"]
        actualHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
        subtaskCount:
          type: integer
        completedSubtaskCount:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, priority, creatorId, creatorName, tags, subtaskCount, completedSubtaskCount, commentCount, createdAt, updatedAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: ["integer", "null"]
        parentTaskId:
          type: ["integer", "null"]
        assigneeId:
          type: ["integer", "null"]
        creatorId:
          type: integer
        dueDate:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        estimatedHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
      required: [title, creatorId]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        estimatedHours:
          type: ["number", "null"]
        actualHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
    TaskComment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, taskId, authorId, authorName, content, createdAt, updatedAt]
    CreateTaskCommentBody:
      type: object
      properties:
        content:
          type: string
        authorId:
          type: integer
      required: [content, authorId]
    AddWatcherBody:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
...