# lib/api-zod/src/generated/types/addWatcherBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clipup Project Management & Messaging API
 * OpenAPI spec version: 0.2.0
 */

export interface AddWatcherBody {
  userId: number;
}